atena/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-crispy-forms==1.7.0
django-localflavor==2.0
idna==2.6
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2017.3
requests==2.18.4
urllib3==1.22
